Updated benchmark

+++ src/benchmark_handler.py
@@ -67,8 +67,9 @@ def _init_bin(bin_path):
     bin_path = os.environ.get('LAMBDA_TASK_ROOT', bin_path)
     cur_file = os.path.join(bin_path, 'storage_bench')
     new_file = os.path.join('/tmp/bin', 'storage_bench')
-    shutil.copy2(cur_file, new_file)
-    os.chmod(new_file, 0o775)
+    if not os.path.isfile(new_file):
+        shutil.copy2(cur_file, new_file)
+        os.chmod(new_file, 0o775)
     return new_file
